fixed paropt examples
